Updated version and logging function, added resource caching
The code changes include an update to the development constants, specifically the version number. The logging function has been modified to only log when in development mode. A new feature for caching resource data by city ID has been introduced. This includes a mechanism to clear the cache when the city changes. Additionally, a new functionality for counting resources in a city has been implemented. Lastly, some minor adjustments have been made to improve readability and organization of the code.