Add italian localization (#2)
* Add italian localization

* Updated version, fixed localization issues and added Italian language support

The software version has been updated. A duplicate entry in the Japanese localization file was removed, and new entries for Italian translations were added. The README file was also updated to reflect these changes, including the addition of Italian language support and a note about fixing semicolon issues in the localization file.